单独拎出来 collector keeper
